// apps/server/src/modules/search/interfaces/search.interface.d.ts
export interface SearchResult {
    rank: number;
    score: number;
    patentId: string;
    title: string;
    abstract: string;
    claims?: string | undefined;
    ipc: string;
    country?: string | undefined;
    owner?: string | undefined;
    publicationDate?: string | undefined;
    section?: string | undefined;
    vectorId?: string | undefined;
}